Refactor : 장소 카테고리 선택 부분 MVVM 모델 적용
장소 카테고리 선택 부분 MVVM 모델 적용하여 Refactoring 함
